refactor: move stack frame creation to utils
